pyproject.toml
[tool.poetry]
name = "django-api-escola-testes"
version = "0.1.0"
description = ""
authors = ["ledsouza <leandro.souza.159@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "^3.8.1"
Django = "^5.0.6"
djangorestframework = "^3.15.1"
Faker = "^25.5.0"
python-dateutil = "^2.9.0.post0"
pytz = "^2024.1"
six = "^1.16.0"
sqlparse = "^0.5.0"
text-unidecode = "^1.3"
validate-docbr = "^1.10.0"
django-cors-headers = "^4.3.1"
pillow = "^10.3.0"
django-redis = "^5.4.0"
redis = "^5.0.5"
hiredis = "^2.3.2"
python-gettext = "^5.0"
django-admin-honeypot = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
